+++ utils/strapi/queries.ts
@@ -23,7 +23,6 @@ import { PROGRAMS_FILTER } from "@/utils/strapi/sections/ProgramsFilter";
 import { PROMO_LINK_LIST } from "@/utils/strapi/sections/PromoLinkList";
 import { RICH_TEXT_IMAGE } from "@/utils/strapi/sections/RichTextImage";
 import { RICH_TEXT_VIDEO } from "@/utils/strapi/sections/RichTextVideo";
-import { ROCKSTAR_INFO } from "@/utils/strapi/sections/RockstarInfo";
 import { ROCKSTARINFO_LIST } from "@/utils/strapi/sections/RockstarInfloList";
 import { STATISTICS_CARD_LIST } from "@/utils/strapi/sections/StatisticsCardList";
 import { TEXT_CONTENT } from "@/utils/strapi/sections/TextContent";
@@ -51,8 +50,7 @@ import type { ProgramsFilterSection } from "@/utils/strapi/sections/ProgramsFilt
 import type { PromoLinkListSection } from "@/utils/strapi/sections/PromoLinkList";
 import type { RichTextImageSection } from "@/utils/strapi/sections/RichTextImage";
 import type { RichTextVideoSection } from "@/utils/strapi/sections/RichTextVideo";
-import { RockstarInfoSection } from "@/utils/strapi/sections/RockstarInfo";
-import { RockstarInfoListSection } from "@/utils/strapi/sections/RockstarInfloList";
+import type { RockstarInfoListSection } from "@/utils/strapi/sections/RockstarInfloList";
 import type { StatisticsCardListSection } from "@/utils/strapi/sections/StatisticsCardList";
 import type { TextContentSection } from "@/utils/strapi/sections/TextContent";
 
@@ -81,7 +79,6 @@ export type ComponentSection =
   | PromoLinkListSection
   | RichTextImageSection
   | RichTextVideoSection
-  | RockstarInfoSection
   | RockstarInfoListSection
   | StatisticsCardListSection
   | TextContentSection
@@ -112,7 +109,6 @@ export const SECTIONS = `
   ${PROMO_LINK_LIST}
   ${RICH_TEXT_IMAGE}
   ${RICH_TEXT_VIDEO}
-  ${ROCKSTAR_INFO}
   ${ROCKSTARINFO_LIST}
   ${STATISTICS_CARD_LIST}
   ${TEXT_CONTENT}
