+++ utils/strapi/sections/RockstarInfloList.ts
@@ -1,4 +1,18 @@
-import { RockstarInfoSection } from "./RockstarInfo";
+import { StrapiImage } from "@/types/strapi/common";
+
+export type RockstarInfoSection = {
+  type: "ComponentSectionsRockstarInfo";
+  name: string;
+  image: StrapiImage;
+  campus: {
+    data: {
+      attributes: {
+        name: string;
+      }
+    }
+  };
+  detail: string;
+}
 
 export type RockstarInfoListSection = {
   type: "ComponentSectionsLinkList";
